Expose getting EndpointData on adapter and session
All of this "get endpoint data" stuff is great, but it's no good if the
user can't ask "hey - what are we working with here". It's safe enough
to not cache this data on the session or adapter objects, because the
source data is all cached anyway. That way calling the method with
different filters will always return the correct data.

Change-Id: I1feb505b2d838bd4b6505336c23b68041a938008