Add 'none' identity plugin
It does not accept any arguments and sets the token to 'notused'.
It does not have any endpoint/url associated,
and thus must be used together with adapter.Adapter.endpoint_override to
instantiate a session for client to a service that is deployed in
noauth/standalone mode.

Unfortunately the 'noauth' name is already taken by
cinderclient.contrib.noauth.

Change-Id: I2f7175ea095aac774e7c5889880fcff15586559c