Refactor volume mock urls in identity_common tests
The large set of copy-pasta test urls make it hard to follow what is
being mocked and what should be returned. Put in a class to manage it
with clearly named things.

This exposes a bug, but fixing the bug is hairy. Comment out that bit
with a note to fix it. The fix comes in the next patch, but is hopefully
clearer to read with the test refactor decoupled.

Change-Id: I1435d2dc16db7b1a7f5dfe76562de56ba8450531