Make Discover.version_data accept null max_version
Make Discover.version_data a little more lenient in the JSON data it
accepts with respect to "max_version" and "version".  Previously, if the
"max_version" field was present but empty, the "version" field would be
ignored, even if populated with a proper version number.  This goes
against the intent of the API.  This change causes version_data to
ignore a blank "max_version" if "version" is populated.

This change also introduces a unit test harness for the various
permutations of version_data translating input version data to
min/max_microversion, including cases for the above.

Change-Id: I1cb4ebdd57a6eadc1a16efc0fc1ceeda640f26c2
Closes-Bug: #1703438