Make discover._version_between more consistent
Partly in response to [1], and partly for the sake of making its API
consistent and understandable, discover._version_between now allows
un-normalized input, and responds sensibly to unspecified upper & lower
bounds.

Previously the method treated None as follows:

min_version  max_version  behavior
None         None         Always False
not None     None         ValueError
not None     not None     Compare min to candidate
None         not None     Compare max to candidate

That's whack, yo.  Took me like ten minutes just to come up with that
chart.

This change allows us to explain clearly what happens when upper or
lower bounds are None: it translates naturally to "there is no bound".

Given that, the first line of the chart (None/None) now always returns
True, which is a behavior change the callers need to be updated to
expect.  This is an acceptable sacrifice for making this method sane.

[1] https://review.openstack.org/#/c/483604/9/keystoneauth1/discover.py@306

Change-Id: I09b6cf692c8bfa290b73c8d7498bc12a5e91d690