Allow discovery URLs to have trailing slashes
The _get_discovery_url_choices generator works by taking a starting URL,
splitting it on '/', and working through the parts trying to get a
matching discovery document from it. It makes assumptions about what the
URL might look like: it might have a project ID on the end of it, and a
version before that. If the starting URL has a trailing '/', splitting
the URL results in an empty string at the end of the list of parts,
which is then treated as a version. The real version is left on the URL
while the generator assumes it has already trimmed the URL down to an
unversioned endpoint. If that version does not match the version we're
seeking, the resulting discovery document will be mismatched and the
generator will fail to yield the right endpoint.

This patch normalizes the starting URL by removing the trailing '/', if
there is one. This way every part of the split URL will be meaningful.

Closes-bug: #1709658

Change-Id: I28c48f78d6f07804d6ea228f163dd37b0fcfcd58