Remove tox_install.sh and align with constraints consumption
As part of removing reliance on the old and deprecated zuul-cloner, we
need to shift constraints declaration to the deps line. This means we
unfortunately have to duplicate the extras declarations into
test-requirements - because otherwise the contraints for keystoneauth
conflicts with the installation of itself.

(cherry picked from commit 8ece7fcf2d5eaa40b7ecf40e81cc16c09a599535)

Change-Id: If162dde93423a495064c4ae1a5dad4e7eb8f9577