Remove tox_install.sh
This patch removes a legacy tool for installing tox which started
breaking stable branches based on openstack-requirements. This is a
handcrafted version of the following patch:

  If162dde93423a495064c4ae1a5dad4e7eb8f9577

This wasn't a direct cherry-pick due to the conflicts with refactoring
to use doc/requirements.txt. This change also includes support for
switching keystoneauth to use stestr. This is because requirements for
stable/pike include os-testr 1.0.0 which uses stestr under the
covers.

Change-Id: I077325d9b8ea8b99a8b5a8ae3de6aef20c7e4f43