Add OTP to v3OIDCpassword plugin
Problem description
===================
When using OpenStack in federated environments, to enhance user's
authentication, some Identity Providers allow them (users) to use an OTP
(One Time Password) method. This requires users to send a generated code
together with his/her username and password any time he/she is
authenticating in the Identity Provider. It can be applied to both Id
and Access tokens authentication flows.

If the user is using the OTP, he/she will not be able to use the
OpenStack CLI, because the user needs to send an OTP value to the
Identity Provider together with the username and password.

Proposal
========
To allow users that have OTP enabled in their Identity Provider to use
the OpenStack CLI with their credentials, we propose to add a new
optional configuration in the v3OIDCpassword plugin to allow users to
enter their OTP Code.

The new configuration will contain the key for the OTP code that the
user's Identity Provider expects in its Rest API. As the OTP is not
defined by the OpenID Connect protocol (it defines just "credentials"
that is a set of attributes that authorize the user, like username,
password, OTP, etc); we defined the property as the expected key to the
OTP in the IdP's Rest API. For example, in KeyCloak, the OTP key in its
Access Token API is called "totp". Therefore, if the user configures
the new property, then the CLI will request the OTP code when necessary.

We do not create a property to set the OTP code specifically,
because the OTP changes every time, so is not so practical to users to
reconfiguring the property every time the OTP code expires.

Change-Id: Ibd27470c9250000d24cf085ccf6b0c31c782c21e