Re-add python3.5 testing
keystoneauth is a little special in the world, in that it's an user
facing library and takes pride in staying compatible. As such it
doesn't change much, so the likelihood that patches will land that
break python3.5 is really low in the first place.

Although the overall openstack python3 support goal puts a min of
python3.6 - which is a great choice, keystoneauth has a big user,
OpenStackSDK, that is keeping python3.5 because OpenStackSDK has
a big user - Zuul, that still supports python3.5.

To be friendly to everyone, keep the py35 unit tests just to make
sure nobody lands f-strings or anything.

Once Zuul drops python3.5 we can circle back around and drop it
here too.

Remove openstackdocstheme from test-requirements.txt. It only
needs to be in doc/requirements.txt.

Change-Id: If380bbf77b757655b1dc55322f4636ef83986fe5