Clarify token_endpoint.Token usecases
This authentication plugin isn't grouped with the other authentication
plugins, but it can be really useful for people writing proxies or
implementing service-to-service calls.

This commit just highlights the usefulness of this plugin with some
example usecases and a code snippet. The current documentation phrases
it such that it's only useful for testing.

Change-Id: I1de4959ccde3fdf8141d8f4949c73542c2200483