Bump bandit
Address incompatibility with Python 3.8, which happens due to the older
bandits use of the private '_ast' module instead of 'ast' [1] and the
deprecated 'Num' having been removed from the former. The bump requires
an additional nosec for a line that is misidentified as a hardcoded
password.

A note about requirements ordering is removed as it hasn't been relevant
since pip 20.3 introduced the new dependency resolver.

[1] https://github.com/PyCQA/bandit/commit/09b0207e2b1a2e17c0f68aee401333abeb5eef66

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Change-Id: Icaaa3a2e24429bba5cf70c04062cfa5820c8a1bf