typing: Annotate keystoneauth1.access.service_providers
This is mostly simple...except for the fact that we're adding TypedDict
representations of all the token response for APIs v2 and v3 going back
to Diablo [1] :) Fortunately these are only used for type checking
(i.e. not at runtime) so if they're incorrect then we will simply
confuse people. They should be very close though.

[1] https://github.com/openstack/keystone/blob/diablo-eol/keystone/logic/types/auth.py#L371-L416

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Change-Id: I9852abf477719253cd38c24f03f3ddcc8b078c11