typing: Annotate keystoneauth1.loading.adapter
We're getting fancy. Unlike most other loaders, this one doesn't return
an identity plugin. Rather, it returns an Adapter. It's probably
reasonable to question whether these need to derive from the same loader
but for now, they do, so we work around this by making 'BaseLoader' a
generic and renaming it to '_BaseLoader', with a concrete implementation
called 'BaseLoader' that most plugin loaders can derive from.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Change-Id: I810822438e738da42e987be82113a9c23c05e87f