typing: Make loaders broadly generic
Yet more fanciness. A Loader class always has a 1:1 relationship with a
given Plugin class. The way to indicate this in the world of typing is
via a generic. Start doing this, allowing us to remove a whole load of
casts (including subclass methods that exist purely to allow us to cast
the return type) and generally tidy everything up a little.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Change-Id: Ic77def3da3453614944589d59a86ec9fef44c955