typing: Deprecate use of AuthConstructor
The AuthConstructor class and its use of '_method_parameters' is one of
the few remaining bit of magic we have left in the codebase after the
typing effort. Remove it in favour of static configuration.

We also fix typing issues with the 'KerberosMethod', 'PasswordMethod'
'TOTPMethod', and 'ApplicationCredentialMethod' classes, where all were
indicating that some configuration parameters were required when in
reality they are optional.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Change-Id: I9fcaea274ca1b6264df93750c5623eea89f11b09