Remove client_id from OIDC payload
Okta does not allow client_id to be in payload and auth header.
Credentials should either be in the auth header, POST body, or
a client_assertation in the POST body. If the client_secret is
set, it will be with client_id in the auth headers. If it is not,
client_id will be added to the payload. This change prevents
client_id from appearing more than once.

Closes-bug: 2105891
Change-Id: I9a7590528aacf46e973e68348775444a9d9bacba