Fix bug in OIDC Device Authorization
There are two bugs in the current device flow using keystoneauth:

1. An issue where an error occurs if `client_secret` is not specified.
In the device flow, authentication should be possible without specifying
`client_secret`[1] but the current implementation requires it.
This bug was introduced by the following patch:
- https://review.opendev.org/c/openstack/keystoneauth/+/932989

2. An issue with the polling process from the client in the device flow
In the device flow, polling is used to verify that the user has
authenticated. There is an issue with the process that checks for
incomplete polling, causing the process to halt with an error:
"Unrecognized schema in response body. (HTTP 400)"
This bug was introduced by the following patch update:
https://review.opendev.org/c/openstack/keystoneauth/+/930769

[1] https://datatracker.ietf.org/doc/html/rfc8628#section-5.6

Closes-bug: #2107373
Change-Id: I6e44e70394d51c9fca06d7f511d1ffe0d9b2c025